pyproject.toml
[project]
name = "deberta-finetune"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pandas>=3.0.0",
    "huggingface-hub>=0.24.0",
    "python-dotenv>=1.0.1",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "torch>=2.2.0",
    "transformers>=4.41.0",
    "tqdm>=4.66.0",
    "safetensors>=0.4.0",
]
